feat: refactor trip state management to use useTripStateAction and improve type safety
